Update and optimize implementation

+++ inc/Helper_functions.hpp
@@ -5,7 +5,7 @@
 
 double calculateRMSE(const std::vector<double>& predictions);
 double calculateMAE(const std::vector<double>& predictions);
-std::vector<int> createRoundedExponentialVector(const double initial_value, const double final_value, const int number_of_elements); // Added semicolon here
+std::vector<long> createRoundedExponentialVector(const double initial_value, const double final_value, const int number_of_elements); // Added semicolon here
 
 
 #endif // HELPER_FUNCTIONS_H
